Make kill with fire script more resilient
[#134349577]

+++ jobs/rabbitmq-server/templates/kill-with-fire.bash
@@ -2,8 +2,16 @@
 
 SHUTDOWN_LOG="${LOG_DIR:-/var/vcap/sys/log/rabbitmq-server}/shutdown_stdout.log"
 
-PID_FILE="${1:?First argument must be a valid PID file}"
-RUNNING_PID=$(cat "$PID_FILE")
+PID_FILE="${1}"
+if [ -z "$PID_FILE" ]; then
+  echo "First argument must be a valid PID file" >> "${SHUTDOWN_LOG}"
+  exit 1
+fi
+
+if ! RUNNING_PID=$(cat "$PID_FILE") ; then
+  echo "PID file did not exist, continuing." >> "${SHUTDOWN_LOG}"
+  exit 0
+fi
 
 if kill -0 "$RUNNING_PID"; then
   echo "We found a rabbitmq-server process during monit stop and we had to kill it" >> "$SHUTDOWN_LOG"
